Don't initialize HandshakeRecordReader buf before reading into it.
This effectively saves a "copy" (of zeros, over the caller_buf directly
before it gets real data). Also restate the rant against ReadBuf::take
-- it should definitely not be called `take` since it's
similar-but-subtly-different from other `take` methods. But it's still
useful for limited reads.

This required more unsafe code. This is a pattern--to hit maximal
performance (avoid copies/pointless initializations), I have to write
unsafe code. One additional advantage to a ReadBuf::take-like API that
updated the "inner" ReadBuf would be that maximally-performant
application code would not need to be unsafe.